Use RuntimeGenericSurface
